anvil_polkadot::substrate_node::service::storage

Type Alias SystemAccountInfo

pub type SystemAccountInfo = AccountInfo<Nonce, AccountData<Balance>>;

Aliased Type

pub struct SystemAccountInfo {
    pub nonce: u32,
    pub consumers: u32,
    pub providers: u32,
    pub sufficients: u32,
    pub data: AccountData<u128>,
}

Fields
nonce: u32

The number of transactions this account has sent.

consumers: u32

The number of other modules that currently depend on this account’s existence. The account cannot be reaped until this is zero.

providers: u32

The number of other modules that allow this account to exist. The account may not be reaped until this and sufficients are both zero.

sufficients: u32

The number of modules that allow this account to exist for their own purposes only. The account may not be reaped until this and providers are both zero.

data: AccountData<u128>

The additional data that belongs to this account. Used to store the balance(s) in a lot of chains.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 80 bytes